requirements.txt
python-dotenv
transformers
langchain
openai
Pillow